+++ src/app/events/Ready.ts
@@ -29,7 +29,8 @@ export default class implements DiscordEvent {
 			// should not wait
 			POAPService.runAutoEndSetup(client, constants.PLATFORM_TYPE_DISCORD).catch(Log.error);
 			POAPService.runAutoEndSetup(client, constants.PLATFORM_TYPE_TWITTER).catch(Log.error);
-			await POAPService.clearExpiredPOAPs();
+			await POAPService.clearExpiredPOAPs().catch(Log.error);
+			POAPService.setupPOAPCleanupCronJob();
 			
 			Log.info(`${constants.APP_NAME} is ready!`);
 		} catch (e) {