+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 ## 2.7.0-RELEASE (2022-01-18)
 
 1. Add more stability to opt-in messages
-2. Add account unlink/status commands
+2. Add account commands
+   - /account link (renamed from /account verify)
+   - /account list
+   - /account unlink
 3. Restrict DEGEN to certain discord servers
 4. Upgrade discord.js -> 13.6.0, upgrade slash-create -> 5.0.3
 5. Introduce basic `/claim` command and prompt user for opt-in on slash command
+6. 
 
 ## 2.6.2-RELEASE (2022-01-13)